requirements.txt
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
PyYAML==5.1.2
SQLAlchemy==1.3.11
Werkzeug==0.16.0
